Changed: separate resources loading from using
