+++ src/viewport.c
@@ -62,8 +62,8 @@ static landscape_tile_t *landscape_tile;
 static void
 draw_map_tile(int x, int y, int mask, int sprite, frame_t *frame)
 {
-	sprite_t *spr = gfx_get_data_object(sprite, NULL);
-	sprite_t *msk = gfx_get_data_object(mask, NULL);
+	sprite_t *spr = data_get_object(sprite, NULL);
+	sprite_t *msk = data_get_object(mask, NULL);
 	sdl_draw_masked_sprite(spr, x, y, msk, NULL, frame);
 }
 
@@ -604,21 +604,21 @@ draw_paths_and_borders(viewport_t *viewport, frame_t *frame)
 static void
 draw_game_sprite(int x, int y, int index, frame_t *frame)
 {
-	void *sprite = gfx_get_data_object(DATA_GAME_OBJECT_BASE + index, NULL);
+	void *sprite = data_get_object(DATA_GAME_OBJECT_BASE + index, NULL);
 	sdl_draw_transp_sprite(sprite, x, y, 1, 0, 0, frame);
 }
 
 static void
 draw_serf(int x, int y, int color, int head, int body, frame_t *frame)
 {
-	sprite_t *s_arms = gfx_get_data_object(DATA_SERF_ARMS_BASE + body, NULL);
-	sprite_t *s_torso = gfx_get_data_object(DATA_SERF_TORSO_BASE + body, NULL);
+	sprite_t *s_arms = data_get_object(DATA_SERF_ARMS_BASE + body, NULL);
+	sprite_t *s_torso = data_get_object(DATA_SERF_TORSO_BASE + body, NULL);
 
 	sdl_draw_transp_sprite(s_arms, x, y, 1, 0, 0, frame);
 	sdl_draw_transp_sprite(s_torso, x, y, 1, 0, color, frame);
 
 	if (head >= 0) {
-		sprite_t *s_head = gfx_get_data_object(DATA_SERF_HEAD_BASE + head, NULL);
+		sprite_t *s_head = data_get_object(DATA_SERF_HEAD_BASE + head, NULL);
 		x += s_arms->b_x;
 		y += s_arms->b_y;
 		sdl_draw_transp_sprite(s_head, x, y, 1, 0, 0, frame);
@@ -628,8 +628,8 @@ draw_serf(int x, int y, int color, int head, int body, frame_t *frame)
 static void
 draw_shadow_and_building_sprite(int x, int y, int index, frame_t *frame)
 {
-	void *shadow = gfx_get_data_object(DATA_MAP_SHADOW_BASE + index, NULL);
-	void *building = gfx_get_data_object(DATA_MAP_OBJECT_BASE + index, NULL);
+	void *shadow = data_get_object(DATA_MAP_SHADOW_BASE + index, NULL);
+	void *building = data_get_object(DATA_MAP_OBJECT_BASE + index, NULL);
 
 	sdl_draw_overlay_sprite(shadow, x, y, 0, frame);
 	sdl_draw_transp_sprite(building, x, y, 1, 0, 0, frame);
@@ -638,8 +638,8 @@ draw_shadow_and_building_sprite(int x, int y, int index, frame_t *frame)
 static void
 draw_shadow_and_building_unfinished(int x, int y, int index, int progress, frame_t *frame)
 {
-	sprite_t *shadow = gfx_get_data_object(DATA_MAP_SHADOW_BASE + index, NULL);
-	sprite_t *building = gfx_get_data_object(DATA_MAP_OBJECT_BASE + index, NULL);
+	sprite_t *shadow = data_get_object(DATA_MAP_SHADOW_BASE + index, NULL);
+	sprite_t *building = data_get_object(DATA_MAP_OBJECT_BASE + index, NULL);
 
 	int h = ((building->h * progress) >> 16) + 1;
 	int y_off = building->h - h;
@@ -1159,17 +1159,17 @@ static void
 draw_water_waves(map_pos_t pos, int x, int y, frame_t *frame)
 {
 	int sprite = DATA_MAP_WAVES_BASE + (((pos ^ 5) + (game.tick >> 3)) & 0xf);
-	sprite_t *s = gfx_get_data_object(sprite, NULL);
+	sprite_t *s = data_get_object(sprite, NULL);
 
 	if (MAP_TYPE_DOWN(pos) < 4 && MAP_TYPE_UP(pos) < 4) {
 		sdl_draw_waves_sprite(s, NULL, x - 16, y, 0, frame);
 	} else if (MAP_TYPE_DOWN(pos) < 4) {
 		int mask = DATA_MAP_MASK_DOWN_BASE + 40;
-		sprite_t *m = gfx_get_data_object(mask, NULL);
+		sprite_t *m = data_get_object(mask, NULL);
 		sdl_draw_waves_sprite(s, m, x, y, 16, frame);
 	} else {
 		int mask = DATA_MAP_MASK_UP_BASE + 40;
-		sprite_t *m = gfx_get_data_object(mask, NULL);
+		sprite_t *m = data_get_object(mask, NULL);
 		sdl_draw_waves_sprite(s, m, x - 16, y, 0, frame);
 	}
 }
@@ -1325,7 +1325,7 @@ draw_row_serf(int x, int y, int shadow, int color, int body, frame_t *frame)
 
 	/* Shadow */
 	if (shadow) {
-		sprite_t *sh = gfx_get_data_object(DATA_SERF_SHADOW, NULL);
+		sprite_t *sh = data_get_object(DATA_SERF_SHADOW, NULL);
 		sdl_draw_overlay_sprite(sh, x, y, 0, frame);
 	}
